desenvolvimento-full-stack-avancado/strawberry_api/requirements.txt
aiosqlite==0.19.0
fastapi==0.97.0
SQLAlchemy==2.0.16
strawberry-graphql==0.186.1
uvicorn==0.22.0
